Rename files and add types
