+++ apps/www/src/routes/portal/brukere/+page.svelte
@@ -4,7 +4,7 @@
 	import AddUserModal from '$lib/components/portal/AddUserModal.svelte';
 	import { mailto } from '$lib/utils';
 	import { formatDate } from '$lib/date';
-	import UserCard from '$lib/components/Usercards/UserCard.svelte';
+	import UserCard from '$lib/components/cards/UserCard.svelte';
 
 	let { data } = $props();
